Alteration of session ID system
Implemented new stack-based session ID system(the old one was based on a
combination of client's IP and port). Stack allows up to 65535 session
id's(which is the maximum port count). Every session pops a session id
from the stack upon the connection, and returns it to the stack upon
disconnection.

Also, a new [] operator overrider function added to the class. You can
now use Server<MyType>[15] to access the session with session id of 15.

Some minor functions added in order to provide manual locking mechanism
for session array's lock(can be needed in some cases), and added a
function for returning session map.